Add `Expression` abstraction (chapter 12)
This commit:
- introduces the `Expression` interface
- adds a simple implementation for `Money.plus`
- adds a *fake* implementation for `Bank.reduce`.

An `Expression` represents the result of arithmetic operations like
`plus` on a `Money` object. It should be able to represent the result
of operations involving different currencies, e.g. "$5 + 3 CHF" would
be the expression representing the addition of 5 Dollars and 3 Francs.

Expressions have to be reduced using `Bank.reduce`
before they can be treated like a simple `Money` objects again.
When a `Bank` reduces an `Expression`, all currencies in the expression get
converted to a single currency and are then added together to return a
`Money` value.